\name{sequence_leading}
\alias{sequence_leading}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Sequence of Possible Leading Digits
}
\description{
Returns a vector containing all possible significant digits for a given number of places.
}
\usage{
sequence_leading(number = 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{number}{An integer determining the number of first digits to be returned, i.e. 1 for 1:9, 2 for 10:99 etc.}
}
\value{
Returns an integer vector.
}

\author{
Dieter William Joenssen \email{Dieter.Joenssen@TU-Ilmenau.de}
}
