\name{rbenf}
\alias{rbenf}

\title{
Random Sample Satisfying Benford's Law
}
\description{
Returns a random sample with length \code{n} satisfying Benford's law.
}
\usage{
rbenf(n)
}
\arguments{
  \item{n}{Number of observations.
}
}

\value{
Returns a random sample with length \code{n} satisfying Benford's law.
}
\references{
Benford F. The law of anomalous numbers. Proceedings of the American Philosophical Society. 1938;78:551-572.\cr
}
\author{
Dieter William Joenssen \email{Dieter.Joenssen@TU-Ilmenau.de}
}

\seealso{
\code{\link{qbenf}}; \code{\link{pbenf}}
}
\keyword{distribution}
\keyword{datagen}
